@@ -5,10 +5,11 @@
   "dependencies": {
     "bulma": "^0.7.2",
     "node-sass": "^4.10.0",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
+    "react": "^16.7.0-alpha.2",
+    "react-dom": "^16.7.0-alpha.2",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "2.0.3"
+    "react-scripts": "2.0.3",
+    "uuid": "^3.3.2"
   },
   "scripts": {
     "start": "react-scripts start",
